pyproject.toml
[project]
name = "mcp-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.8.1",
    "google-cloud-bigquery>=3.34.0",
    "langchain-community>=0.3.26",
    "langchain-core>=0.3.66",
    "langchain-google-genai>=2.1.5",
    "langchain-mcp-adapters>=0.1.7",
    "langchain-ollama>=0.3.3",
    "langchain-openai>=0.3.25",
    "langgraph>=0.4.10",
    "langgraph-cli[inmem]>=0.3.3",
    "langgraph-prebuilt>=0.2.3",
    "python-dotenv>=1.1.0",
]
